create new component CreateForm: remove component CreateModal, put CreateForm component to props to Modal in AddBookGroup component
